pyproject.toml
[tool.poetry]
name = "arbitrage"
version = "0.1.0"
description = ""
authors = ["Massil Achab <massil.achab@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.0"
requests = "^2.28.1"
pandas = "^1.4.4"
rel = "^0.4.7"

[tool.poetry.dev-dependencies]
websocket-client = "1.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
